Add files via upload
The primary objective of this project is to develop a predictive model that can forecast the outcomes of IPL matches based on historical match data. This model aims to provide insights and predictions for cricket enthusiasts and analysts, enhancing their understanding and enjoyment of the game.

The project involves analyzing the IPL match dataset obtained from Kaggle using Python and Jupyter Notebook. The analysis will utilize libraries such as pandas for data manipulation, and the model will be saved using the pickle library. Streamlit will be used to create an interactive web application that allows users to input match details and receive predictions.